Fix critical errors with other plugins in Rest.php
One other plugin using rest causes this when it's trying to do its own calls:
CRITICAL Uncaught Error: Cannot use object of type Closure as array in /multisite-global-media/src/Rest/Rest.php:76

Added !is_callable to avoid this. I'm not really familiar with this plugin so I'm not sure if there is better fix for this. Also probably something should be done to avoid this mixup in the first place.